+++ src/main/java/com/ya/pf/auditable/payment_method/dto/PaymentMethodDTO.java
@@ -1,5 +1,5 @@
 package com.ya.pf.auditable.payment_method.dto;
 
-public record PaymentMethodDTO(long id, String name, double balance) {
+public record PaymentMethodDTO(long id, String name, float balance) {
 
 }
