+++ src/main/java/com/ya/pf/auditable/transaction/owner_transaction/entity/OwnerTransactionService.java
@@ -4,10 +4,10 @@ import java.util.Date;
 
 public interface OwnerTransactionService {
 
-    void createOwnerTransaction(long supplierId, double amount, Long paymentId, Long billId, Date date);
+    void createOwnerTransaction(long supplierId, float amount, Long paymentId, Long billId, Date date);
 
-    void deleteOwnerTransactionByBillId(long billId, double billAmount);
+    void deleteOwnerTransactionByBillId(long billId, float billAmount);
 
-    void deleteOwnerTransactionByPaymentId(long paymentId, double paymentAmount);
+    void deleteOwnerTransactionByPaymentId(long paymentId, float paymentAmount);
 
 }
